Testing Scourge Dimenion Hideout Decoration
Trying to figure out how to get it so that can have anime girl in main dimension, then another when you click on the scourge hideout decoration.

Discoveries so far:
fv seems to be correlated to scourge dimension (fv = 64)
To move items into the scourge dimension, you need to first have the scourge dimension item listed (in the doodads list) before having everythinig after only being items in the scourge dimension (everything must have an fv of 64). The order of an item is tacked onto the end of the file (so the next decoration you add gets placed at the end of the file, so it seems like order matters for the scourge dimension)

My next idea is to test to see if fv is the layering or exposure or visibility value or something by changing it and seeing if it changes the image